Android: enable lineHeight as a decimal number
Summary:
For some reason android will not accept a decimal
value as an lineHeight even though it's according
to css a valid value.

Change the java-backing class type to float instead
according to this issue:
https://github.com/facebook/react-native/issues/10607

Verification:

Test Plan:
* Run app on android with decimal values. App no longer dies with error-message TypeError: expected dynamic type int64, but has type double.
* IOS version unaffected of change.

Reviewers: meros

Reviewed By: meros

Subscribers: jonas

Differential Revision: http://192.168.5.120.xip.io/D14

+++ ReactAndroid/src/main/java/com/facebook/react/views/text/ReactTextShadowNode.java
@@ -340,7 +340,7 @@ public class ReactTextShadowNode extends LayoutShadowNode {
   protected int mNumberOfLines = UNSET;
   protected int mFontSize = UNSET;
   protected float mFontSizeInput = UNSET;
-  protected int mLineHeightInput = UNSET;
+  protected float mLineHeightInput = UNSET;
   protected int mTextAlign = Gravity.NO_GRAVITY;
   protected int mTextBreakStrategy = (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) ?
       0 : Layout.BREAK_STRATEGY_HIGH_QUALITY;
@@ -443,7 +443,7 @@ public class ReactTextShadowNode extends LayoutShadowNode {
   }
 
   @ReactProp(name = ViewProps.LINE_HEIGHT, defaultInt = UNSET)
-  public void setLineHeight(int lineHeight) {
+  public void setLineHeight(float lineHeight) {
     mLineHeightInput = lineHeight;
     if (lineHeight == UNSET) {
       mLineHeight = Float.NaN;
